+++ src/views/SettingsWindow/SettingsWindow.tsx
@@ -14,6 +14,7 @@ import { SystemTheme } from "../../theme/SystemTheme";
 import { AppSettings } from "../../settings/AppSettings";
 import General from "./pages/General";
 import Appearance from "./pages/Appearance";
+import Editor from "./pages/Editor";
 import Shortcuts from "./pages/Shortcuts";
 import DataStorage from "./pages/DataStorage";
 import License from "./pages/License";
@@ -59,6 +60,9 @@ function SettingsWindow(props: SettingsWindowProps) {
   switch (selectedPage) {
     case SettingsView.appearance:
       page = <Appearance theme={props.theme} appSettings={props.appSettings} onAppSettingsChange={props.onAppSettingsChange} />;
+      break;
+    case SettingsView.editor:
+      page = <Editor appSettings={props.appSettings} onAppSettingsChange={props.onAppSettingsChange} />;
       break;
     case SettingsView.shortcuts:
       page = <Shortcuts />;